hashes: Create write_hash_file and use it in wheel creation
Also move hash_of_file to utils/hashes.py
